requirements.txt
fastapi
uvicorn[standard]
groq
weaviate-client
mysql-connector-python
python-dotenv